Rename relaxer and suggester
