+++ internal/remediation/relax/npm.go
@@ -1,4 +1,4 @@
-package relaxer
+package relax
 
 import (
 	"context"
